day3
rather than shenanigans to read the input in I just shoved it in the
code. It doesn't matter.

I feel like this could be much more efficient than enumerating the
entire list of points and such, and especially since golang doesn't have
set functions which could have made this possibly cleaner, but it ran in
~no time anyways so I don't particularly care :)