day15
a few days late but I was having trouble getting brain to work saturday
night and on the way home tonight I realized what I had done wrong. When
I reached the end of a fork and was rolling back I wasn't moving the
droid back so it was just ending at the first fork end. Oops