day16
part1's code isn't what I ended up using actually because I hacked up
the day16 library so much and didn't change the tests or whatever.
I tried a big "make 100 goroutines and run this all in parallel"
situation before realizing that this isn't super parallelizable and also
the massive shortcut.

I'm ... not a huge fan of the "clever trick" types of things, I'd prefer
it be more about optimizing code or taking advantage of language
features (threading, etc) but ok. I definitely needed a hint on this one
and even that hint took me a while to process, but once I did, obviously
I figured out the massive shortcut :)